+++ client/src/lib/navigation/ModalStack.tsx
@@ -9,6 +9,7 @@ import SessionModal from '../../routes/SessionModal/SessionModal';
 import CreateSessionModal from '../../routes/CreateSessionModal/CreateSessionModal';
 import AddSessionModal from '../../routes/AddSessionModal/AddSessionModal';
 import UpgradeAccountModal from '../../routes/UpgradeAccountModal/UpgradeAccountModal';
+import RequestPublicHostModal from '../../routes/RequestPublicHostModal/RequestPublicHostModal';
 import SessionUnavailableModal from '../../routes/SessionUnavailableModal/SessionUnavailableModal';
 import {COLORS} from '../../../../shared/src/constants/colors';
 import SETTINGS from '../constants/settings';
@@ -154,11 +155,20 @@ const ModalStack = () => {
           component={SessionFeedbackModal}
           options={shortSheetModalScreenOptions}
         />
+
+        <Screen
+          name={'UpgradeAccountModal'}
+          component={UpgradeAccountModal}
+          options={shortSheetModalScreenOptions}
+        />
       </Group>
 
       <Group screenOptions={cardModalScreenOptions}>
         <Screen name={'AddSessionModal'} component={AddSessionModal} />
-        <Screen name={'UpgradeAccountModal'} component={UpgradeAccountModal} />
+        <Screen
+          name={'RequestPublicHostModal'}
+          component={RequestPublicHostModal}
+        />
       </Group>
     </Navigator>
   );
